backend/requirements.txt
flask==3.1.3
flask-cors==6.0.2
openai-whisper==20250625
sentence-transformers==5.2.3
scikit-learn==1.8.0
pandas==3.0.1
numpy==2.4.2
edge-tts==7.2.7
gunicorn==23.0.0